enhance brushExtents() based on #53
