+++ src/brush/1d/brushExtents.js
@@ -1,6 +1,8 @@
 import { select } from 'd3-selection';
 import { brushSelection } from 'd3-brush';
 
+import invertByScale from '../invertByScale';
+
 const brushExtents = (state, config, pc) => extents => {
   const { brushes, brushNodes } = state;
 
@@ -9,7 +11,17 @@ const brushExtents = (state, config, pc) => extents => {
       const brush = brushes[cur];
       //todo: brush check
       if (brush !== undefined && brushSelection(brushNodes[cur]) !== null) {
-        acc[cur] = brush.extent();
+        const raw = brushSelection(brushNodes[cur]);
+        const yScale = config.dimensions[cur].yscale;
+        const scaled = invertByScale(raw, yScale);
+
+        acc[cur] = {
+          extent: brush.extent(),
+          selection: {
+            raw,
+            scaled,
+          },
+        };
       }
 
       return acc;
